+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Take host environment variables to configure a http / https proxy for `docker-inside-setup`
 
 ## [0.3.14] - 2018-07-07
 ### Added